Refactor volume into a separate modules
